Revert "Update fork"

+++ demos/python_demos/speech_recognition_demo/ctcdecode-numpy/ctcdecode_numpy/yoklm/vocabulary.hpp
@@ -18,9 +18,9 @@ namespace yoklm {
 typedef uint64_t WordHash;
 
 struct VocabularyConfig {
-  size_t num_words;  // the number of tokens in vocabulary, initialized with counts[0]
+  int num_words;  // the number of tokens in vocabulary, initialized with counts[0]
   MemorySectionArray<WordHash> word_hashes;  // sorted
-  size_t num_word_strings;  // the number of tokens presented as strings. It can be 0 or num_words
+  int num_word_strings;  // the number of tokens presented as strings. It can be 0 or num_words
   MemorySectionArray<char> word_strings;
 };
